Extract shared methods for image processing and logging across Freelex and Signbank
