+++ build-assets-from-freelex.py
@@ -2,16 +2,11 @@
 import os
 import re
 import shutil
-import sys
 import xml.etree.ElementTree as ET
 from optparse import OptionParser
 
 import freelex
-
-
-def print_run_msg(msg):
-    print(" - Running: " + msg)
-
+from log import print_run_msg
 
 parser = OptionParser()
 parser.add_option("-c", action="store_true", dest="cleanup",
@@ -50,49 +45,11 @@ freelex.write_sqlitefile()
 print("Step 6: Merge images together into one folder")
 freelex.copy_images_to_one_folder()
 
-print("Step 6a: Generate search thumbnails")
-# Create thumbnails used on search screens
-for path, dirs, files in os.walk("assets/"):
-    for filename in files:
-        # Some images have a 1px border that looks bad in search results
-        # Not all do - but we can safely trim 1px off all images
-        shave_cmd = "mogrify -shave 1x1 assets/" + filename
-        print_run_msg(shave_cmd)
-        os.system(shave_cmd)
-
-        # Then we make thumbnails of the border-free images
-        create_thumbnail_cmd = "convert -resize x92 assets/" + \
-            filename + " assets/50." + filename
-        print_run_msg(create_thumbnail_cmd)
-        os.system(create_thumbnail_cmd)
-
-print("Step 6p: Shrink images for distribution")
-# In order to keep the app size small, we need to run a series
-# of compressions over the images
-
-# Resize images larger than 600x600 down using mogrify from imagemagick
-for path, dirs, files in os.walk("assets/"):
-    for filename in files:
-        cmd = "mogrify -resize '600x600>' assets/" + filename
-        print_run_msg(cmd)
-        os.system(cmd)
-
-# Convert all images to 4 colour depth
-for path, dirs, files in os.walk("assets/"):
-    for filename in files:
-        recolor_cmd = "convert -colors 4 assets/" + filename + " assets/" + filename
-        print_run_msg(recolor_cmd)
-        os.system(recolor_cmd)
-
-# Finally, run optipng
-for path, dirs, files in os.walk("assets/"):
-    for filename in files:
-        optipng_cmd = "optipng -quiet assets/" + filename
-        print_run_msg(optipng_cmd)
-        os.system(optipng_cmd)
+print("Step 7: Prepare images for distribution")
+image_processing.process_images(pictures_folder)
 
 if options.cleanup:
-    print("Step 7: Cleanup")
+    print("Step 8: Cleanup")
     os.remove("dnzsl-xmldump.xml")
     os.remove("nzsl.dat")
     os.remove("nzsl.db")
