+++ log.py
@@ -1,0 +1,2 @@
+def print_run_msg(msg):
+    print(" - Running: " + msg)
